Separate the list trim from the render

+++ cpu.c
@@ -1,6 +1,6 @@
 #include "modules.h"
 
-int cpu_perc() {
+float cpu_perc() {
   static long double a[7] = {0};
   long double b[7], sum;
   //if(skip_read) return 0;
@@ -31,7 +31,7 @@ int cpu_perc() {
       return -1;
   }
 
-  return (int)(100 * ((b[0] + b[1] + b[2] + b[5] + b[6]) -
+  return (float)(100 * ((b[0] + b[1] + b[2] + b[5] + b[6]) -
                       (a[0] + a[1] + a[2] + a[5] + a[6])) / sum);
 }
